don't try to serialize size_t into a TokenBuffer
breaks on macos

+++ src/kernel/vrbclient/RegistryClass.cpp
@@ -103,7 +103,7 @@ namespace vrb
 		tb << m_classID;
 		tb << m_isDel;
 
-		tb << m_variables.size();
+		tb << (uint32_t)m_variables.size();
 		for (const auto var : m_variables)
 		{
 			vrb::serialize(tb, *var.second);
@@ -116,9 +116,9 @@ namespace vrb
 		tb >> m_classID;
 		tb >> m_isDel;
 
-		size_t size;
+		uint32_t size;
 		tb >> size;
-		for (size_t i = 0; i < size; i++)
+		for (uint32_t i = 0; i < size; i++)
 		{
 			auto var = createVar("", DataHandle{});
 			vrb::deserialize(tb, *var);
